+++ tests/libtest/lib586.c
@@ -120,7 +120,7 @@ static void *t586_test_fire(void *ptr)
   if(code != CURLE_OK) {
     int i = 0;
     curl_mfprintf(stderr, "perform url '%s' repeat %d failed, curlcode %d\n",
-                  tdata->url, i, (int)code);
+                  tdata->url, i, code);
   }
 
   curl_mprintf("CLEANUP\n");
@@ -236,7 +236,7 @@ test_cleanup:
   scode = curl_share_cleanup(share);
   if(scode != CURLSHE_OK)
     curl_mfprintf(stderr, "curl_share_cleanup failed, code errno %d\n",
-                  (int)scode);
+                  scode);
 
   curl_mprintf("GLOBAL_CLEANUP\n");
   curl_global_cleanup();
